Major changes on Canvas2D and introducing GUI
 - Moved Canvas2D to a new folder location, make it a separate compiled .js file that wll contain everything which is Canvas2D related (so including the GUI lib)
 - Include the first commit of the GUI related files
