+++ Canvas2D/src/GUI/babylon.gui.label.ts
