+++ Canvas2D/src/Shaders/ellipse2d.fragment.fx
