+++ Canvas2D/src/Shaders/ellipse2d.vertex.fx
