+++ Canvas2D/src/Shaders/rect2d.fragment.fx
