+++ Canvas2D/src/Shaders/text2d.fragment.fx
