+++ Canvas2D/src/Shaders/text2d.vertex.fx
